Fix btf_reloc_id_get()
We're using a hashmap to store the ids, in order to avoid storing a
pointer to an integer we just store the integer directly into the hasmap
value by using uint_as_hash_key().

The hashmap saves pointers, then the variable we're going to store
the value in must have the right size. This commit uses uintptr_t
to avoid a stack corruption.