libbpf: Implement API for generating BTF for ebpf objects
This commit implements a new set of "bpf_reloc_info__" functions for generating
a BTF file with the types a given set of eBPF objects need for their CO-RE
relocations. This code reuses all the existing CO-RE logic (candidate lookup,
matching, etc). The workflow is the same as when an eBPF program is being
loaded, but instead of patching the eBPF instruction we save the type involved
in the relocation.

A new struct btf_reloc_info is defined to save the BTF types needed by a set of
eBPF objects. It's created with the bpf_reloc_info__new() function and populated
with bpf_object__reloc_info_gen() for each eBPF object, finally the BTF file is
genereted with bpf_reloc_info__get_btf(). Please take at a look at BTFGen[0] to
get a complete example of how this API can be used.

bpf_object__reloc_info_gen() ends up calling btf_reloc_info_gen_field() that
uses the access spec to add all the types that are needed by a given relocation.
The root type is added, if there are any struct/unions then the member involved
in the relocation is added as well. Reference types are resolved adding all the
referenced types. This function can be called multiple times to add the types
needed for different objects into the same struct btf_reloc_info, this allows
the user to create a BTF file that can be used for multiple eBPF objects.

The bpf_reloc_info__get_btf() generates the BTF file from a given struct
btf_reloc_info. This function first creates a new BTF object and copies all the
types saved in the struct btf_reloc_info there. For structures/unions only the
members involved in a relocation are copied. While adding the types to the new
BTF object, a map is filled with the type IDs on the old and new BTF structures.
This map is then used later on to fix all the IDs in the new BTF object.

Right now only support for field based CO-RE relocations is supported.

[0]: https://github.com/kinvolk/btfgen

Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>
Signed-off-by: Rafael David Tinoco <rafaeldtinoco@gmail.com>
Signed-off-by: Lorenzo Fontana <lorenzo.fontana@elastic.co>