libbpf: Introduce 'btf_custom' to 'bpf_obj_open_opts'
1373ff599556 ("libbpf: Introduce 'btf_custom_path' to 'bpf_obj_open_opts'")
introduced btf_custom_path which allows developers to specify a BTF file
path to be used for CO-RE relocations. This implementation parses and
releases the BTF file for each bpf object. This is not optimal for
applications that load multiple bpf objects.

This commit introduces a new 'btf_custom' option to allow users to
specify directly the btf object instead of the path. This avoids to
parse/release the same BTF file multiple times when the application
loads multiple bpf objects.

Our specific use case is BTFGen[0], there we want to reuse the
same BTF file with multiple bpf objects. In this case passing btf_custom_path
is not only inefficient but it also complicates the implementation as we
want to save pointers of BTF types but they are invalidated after the
bpf object is closed with bpf_object__close().

[0]: https://lore.kernel.org/bpf/20211027203727.208847-1-mauricio@kinvolk.io/

Signed-off-by: Mauricio Vásquez <mauricio@kinvolk.io>
Signed-off-by: Rafael David Tinoco <rafael.tinoco@aquasec.com>
Signed-off-by: Lorenzo Fontana <lorenzo.fontana@elastic.co>